+++ Riot/ViewController/MasterTabBarController.m
@@ -796,7 +796,7 @@
                                                            self->currentAlert = nil;
                                                        }
                                                        
-                                                       [[AppDelegate theDelegate] startGoogleAnalytics];
+                                                       [[AppDelegate theDelegate] startAnalytics];
                                                        
                                                    }]];